create labels command added
